refactor : (이전 리뷰 중) 요금 계산 로직을 Fare 객체로 이관
