+++ src/main/java/nextstep/subway/domain/Fare.java
@@ -1,0 +1,42 @@
+package nextstep.subway.domain;
+
+public class Fare {
+
+    private static final int BASIC_FARE = 1_250;
+
+    private static final int BASIC_FARE_KM = 10;
+
+    private static final int _100WON_PER_5KM_BASED_KM = 50;
+
+    private static final int EXTRA_FARE_BASIC_UNIT = 100;
+
+    private int distance;
+
+    public Fare(int distance) {
+        this.distance = distance;
+    }
+
+    public int calculate() {
+        return BASIC_FARE + calculateOverFare(distance);
+    }
+
+    private int calculateOverFare(int distance) {
+        if (distance <= BASIC_FARE_KM) {
+            return 0;
+        }
+
+        if (distance <= _100WON_PER_5KM_BASED_KM) {
+            return calculateOverFare(BASIC_FARE_KM, distance, 5);
+        }
+
+        int extraFareUnder50Km = calculateOverFare(BASIC_FARE_KM, _100WON_PER_5KM_BASED_KM, 5);
+        int extraFareOver50Km = calculateOverFare(_100WON_PER_5KM_BASED_KM, distance, 8);
+
+        return extraFareUnder50Km + extraFareOver50Km;
+
+    }
+
+    private int calculateOverFare(int from, int to, int divisor) {
+        return (int) ((Math.ceil((to - from - 1) / divisor) + 1) * EXTRA_FARE_BASIC_UNIT);
+    }
+}
